+++ scripts/upload_mappings.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-MAPPINGS_DIR="auto_generated_mappings"
+MAPPINGS_DIR="mappings"
 NUM_CHANGES=$(git status $MAPPINGS_DIR --porcelain | wc -l)
 CURRENT_DATE=$(date +'%m/%d/%y')
 
@@ -9,7 +9,7 @@ if [[ $NUM_CHANGES -gt 0 ]]; then
     git config --global user.name 'Jad Chaar'
     git config --global user.email 'jadchaar@users.noreply.github.com'
     git add $MAPPINGS_DIR
-    git commit -m "$CURRENT_DATE CRON Job: Update auto-generated mappings"
+    git commit -m "$CURRENT_DATE automated CRON job - update mappings"
     git push
 else
     echo "No mapping changes detected. Skipping upload..."